Inheritance and polymorphism Final Commit

+++ src/shapes/Rectangle.java
@@ -2,11 +2,11 @@ package shapes;
 
 public class Rectangle {
 
-    protected int length;
-    protected int width;
+    protected double length;
+    protected double width;
 
 
-    public Rectangle(int length, int width){
+    public Rectangle(double length, double width) {
 
         this.length = length;
 
@@ -14,35 +14,17 @@ public class Rectangle {
     }
 
 
-    public double getArea(){
+    public double getArea() {
 
         return length * width;
 
     }
 
 
-    public double getPerimeter(){
+    public double getPerimeter() {
 
         return 2 * length + 2 * width;
     }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 
 }
